Extracted view-related utility methods from "utils/utils" to "ui/utils" so that the "utils/utils" module does not depend on the view module.
