+++ ui/utils/utils.d.ts
@@ -1,0 +1,12 @@
+﻿declare module "ui/utils" {
+    import view = require("ui/core/view");
+    module ios {
+        /**
+         * Gets actual height of a [UIView](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UIView_Class/) widget.
+         * @param uiView - An instance of UIView.
+         */
+        export function getActualHeight(uiView: UIView): number;
+
+        export function _layoutRootView(rootView: view.View, parentBounds: CGRect): void;
+    }
+}
